+++ typescript/infra/scripts/check-deploy.ts
@@ -8,9 +8,13 @@ import {
   HyperlaneIsmFactory,
   InterchainAccount,
   InterchainAccountChecker,
+  InterchainAccountConfig,
   InterchainQuery,
   InterchainQueryChecker,
+  attachContractsMapAndGetForeignDeployments,
+  hypERC20factories,
 } from '@hyperlane-xyz/sdk';
+import { objFilter } from '@hyperlane-xyz/utils';
 
 import { Contexts } from '../config/contexts.js';
 import { DEPLOYER } from '../config/environments/mainnet3/owners.js';
@@ -25,17 +29,20 @@ import { logViolationDetails } from '../src/utils/violation.js';
 
 import {
   Modules,
-  getAddresses,
   getArgs as getRootArgs,
+  getWarpAddresses,
   withChain,
   withContext,
   withModuleAndFork,
+  withWarpRouteId,
 } from './agent-utils.js';
 import { getEnvironmentConfig, getHyperlaneCore } from './core-utils.js';
 import { getHelloWorldApp } from './helloworld/utils.js';
 
 function getArgs() {
-  return withChain(withModuleAndFork(withContext(getRootArgs())))
+  return withChain(
+    withWarpRouteId(withModuleAndFork(withContext(getRootArgs()))),
+  )
     .boolean('asDeployer')
     .default('asDeployer', false)
     .boolean('govern')
@@ -44,8 +51,16 @@ function getArgs() {
 }
 
 async function check() {
-  const { fork, govern, module, environment, context, chain, asDeployer } =
-    await getArgs();
+  const {
+    fork,
+    govern,
+    module,
+    environment,
+    context,
+    chain,
+    asDeployer,
+    warpRouteId,
+  } = await getArgs();
   const envConfig = getEnvironmentConfig(environment);
   let multiProvider = await envConfig.getMultiProvider();
 
@@ -74,7 +89,15 @@ async function check() {
     multiProvider,
   );
   const routerConfig = core.getRouterConfig(envConfig.owners);
-  const ica = InterchainAccount.fromAddressesMap(chainAddresses, multiProvider);
+  const icaChainAddresses = objFilter(
+    chainAddresses,
+    (chain, addresses): addresses is Record<string, string> =>
+      !!chainAddresses[chain]?.interchainAccountRouter,
+  );
+  const ica = InterchainAccount.fromAddressesMap(
+    icaChainAddresses,
+    multiProvider,
+  );
 
   let governor: HyperlaneAppGovernor<any, any>;
   if (module === Modules.CORE) {
@@ -84,7 +107,7 @@ async function check() {
       envConfig.core,
       ismFactory,
     );
-    governor = new HyperlaneCoreGovernor(checker);
+    governor = new HyperlaneCoreGovernor(checker, ica);
   } else if (module === Modules.INTERCHAIN_GAS_PAYMASTER) {
     const igp = HyperlaneIgp.fromAddressesMap(chainAddresses, multiProvider);
     const checker = new HyperlaneIgpChecker(multiProvider, igp, envConfig.igp);
@@ -93,7 +116,10 @@ async function check() {
     const checker = new InterchainAccountChecker(
       multiProvider,
       ica,
-      routerConfig,
+      objFilter(
+        routerConfig,
+        (chain, _): _ is InterchainAccountConfig => !!icaChainAddresses[chain],
+      ),
     );
     governor = new ProxiedRouterGovernor(checker);
   } else if (module === Modules.INTERCHAIN_QUERY_SYSTEM) {
@@ -123,15 +149,44 @@ async function check() {
     );
     governor = new ProxiedRouterGovernor(checker);
   } else if (module === Modules.WARP) {
-    const config = await getWarpConfig(multiProvider, envConfig);
-    const addresses = getAddresses(environment, Modules.WARP);
+    if (!warpRouteId) {
+      throw new Error('Warp route id required for warp module');
+    }
+    const config = await getWarpConfig(multiProvider, envConfig, warpRouteId);
+    const addresses = getWarpAddresses(warpRouteId);
     const filteredAddresses = Object.keys(addresses) // filter out changes not in config
       .filter((key) => key in config)
       .reduce((obj, key) => {
         obj[key] = addresses[key];
         return obj;
       }, {} as typeof addresses);
-    const app = HypERC20App.fromAddressesMap(filteredAddresses, multiProvider);
+
+    const { contractsMap, foreignDeployments } =
+      attachContractsMapAndGetForeignDeployments(
+        filteredAddresses,
+        hypERC20factories,
+        multiProvider,
+      );
+
+    // log error and return is foreign deployment chain is specifically checked
+    if (
+      (chain && foreignDeployments[chain]) ||
+      (fork && foreignDeployments[fork])
+    ) {
+      console.log(
+        `${
+          chain ?? fork
+        } is non evm and it not compatible with warp checker tooling`,
+      );
+      return;
+    }
+
+    const app = new HypERC20App(
+      contractsMap,
+      multiProvider,
+      undefined,
+      foreignDeployments,
+    );
 
     const checker = new HypERC20Checker(
       multiProvider,