+++ typescript/infra/src/govern/HyperlaneAppGovernor.ts
@@ -41,6 +41,12 @@ export type AnnotatedCallData = CallData & {
   description: string;
 };
 
+export type InferredCall = {
+  type: SubmissionType;
+  chain: ChainName;
+  call: AnnotatedCallData;
+};
+
 export abstract class HyperlaneAppGovernor<
   App extends HyperlaneApp<any>,
   Config extends OwnableConfig,
@@ -161,10 +167,6 @@ export abstract class HyperlaneAppGovernor<
     this.calls[chain].push(call);
   }
 
-  protected popCall(chain: ChainName): AnnotatedCallData | undefined {
-    return this.calls[chain].pop();
-  }
-
   protected async mapViolationsToCalls(): Promise<void> {
     const callObjs = await Promise.all(
       this.checker.violations.map((violation) =>
@@ -184,37 +186,48 @@ export abstract class HyperlaneAppGovernor<
   ): Promise<{ chain: string; call: AnnotatedCallData } | undefined>;
 
   protected async inferCallSubmissionTypes() {
-    await Promise.all(
-      Object.keys(this.calls).map(async (chain) => {
-        try {
-          await Promise.all(
-            this.calls[chain].map(async (call) => {
-              let submissionType = await this.inferCallSubmissionType(
-                chain,
-                call,
-              );
-              if (submissionType === SubmissionType.MANUAL) {
-                submissionType = await this.inferICAEncodedSubmissionType(
-                  chain,
-                  call,
-                );
-              }
-              call.submissionType = submissionType;
-            }),
-          );
-        } catch (error) {
-          console.error(
-            `Error inferring call submission types for chain ${chain}: ${error}`,
-          );
+    const newCalls: ChainMap<AnnotatedCallData[]> = {};
+
+    const pushNewCall = (inferredCall: InferredCall) => {
+      newCalls[inferredCall.chain] = newCalls[inferredCall.chain] || [];
+      newCalls[inferredCall.chain].push({
+        submissionType: inferredCall.type,
+        ...inferredCall.call,
+      });
+    };
+
+    for (const chain of Object.keys(this.calls)) {
+      try {
+        for (const call of this.calls[chain]) {
+          let inferredCall: InferredCall;
+
+          inferredCall = await this.inferCallSubmissionType(chain, call);
+          // If it's a manual call, it means that we're not able to make the call
+          // from a signer or Safe. In this case, we try to infer if it must be sent
+          // from an ICA controlled by a remote owner. This new inferred call will be
+          // unchanged if the call is not an ICA call after all.
+          if (inferredCall.type === SubmissionType.MANUAL) {
+            inferredCall = await this.inferICAEncodedSubmissionType(
+              chain,
+              call,
+            );
+          }
+          pushNewCall(inferredCall);
         }
-      }),
-    );
+      } catch (error) {
+        console.error(
+          `Error inferring call submission types for chain ${chain}: ${error}`,
+        );
+      }
+    }
+
+    this.calls = newCalls;
   }
 
   protected async inferICAEncodedSubmissionType(
     chain: ChainName,
     call: AnnotatedCallData,
-  ): Promise<SubmissionType> {
+  ): Promise<InferredCall> {
     const multiProvider = this.checker.multiProvider;
     const signer = multiProvider.getSigner(chain);
     if (this.interchainAccount) {
@@ -248,7 +261,11 @@ export abstract class HyperlaneAppGovernor<
           config: accountConfig,
         });
         if (!callRemote.to || !callRemote.data) {
-          return SubmissionType.MANUAL;
+          return {
+            type: SubmissionType.MANUAL,
+            chain,
+            call,
+          };
         }
         const encodedCall: AnnotatedCallData = {
           to: callRemote.to,
@@ -256,23 +273,43 @@ export abstract class HyperlaneAppGovernor<
           value: callRemote.value,
           description: `${call.description} - interchain account call from ${origin} to ${chain}`,
         };
-        const subType = await this.inferCallSubmissionType(origin, encodedCall);
+        const { type: subType } = await this.inferCallSubmissionType(
+          origin,
+          encodedCall,
+          (chain: ChainName, submitterAddress: Address) => {
+            // Require the submitter to be the owner of the ICA on the origin chain.
+            return (
+              chain === origin &&
+              eqAddress(bytes32ToAddress(accountConfig.owner), submitterAddress)
+            );
+          },
+        );
         if (subType !== SubmissionType.MANUAL) {
-          this.popCall(chain);
-          this.pushCall(origin, encodedCall);
-          return subType;
+          return {
+            type: subType,
+            chain: origin,
+            call: encodedCall,
+          };
         }
       } else {
         console.log(`Account's owner ${localOwner} is not ICA router`);
       }
     }
-    return SubmissionType.MANUAL;
+    return {
+      type: SubmissionType.MANUAL,
+      chain,
+      call,
+    };
   }
 
   protected async inferCallSubmissionType(
     chain: ChainName,
     call: AnnotatedCallData,
-  ): Promise<SubmissionType> {
+    additionalTxSuccessCriteria?: (
+      chain: ChainName,
+      submitterAddress: Address,
+    ) => boolean,
+  ): Promise<InferredCall> {
     const multiProvider = this.checker.multiProvider;
     const signer = multiProvider.getSigner(chain);
     const signerAddress = await signer.getAddress();
@@ -281,7 +318,32 @@ export abstract class HyperlaneAppGovernor<
       chain: ChainName,
       submitterAddress: Address,
     ): Promise<boolean> => {
+      // The submitter needs to have enough balance to pay for the call.
+      // Surface a warning if the submitter's balance is insufficient, as this
+      // can result in fooling the tooling into thinking otherwise valid submission
+      // types are invalid.
+      if (call.value !== undefined) {
+        const submitterBalance = await multiProvider
+          .getProvider(chain)
+          .getBalance(submitterAddress);
+        if (submitterBalance.lt(call.value)) {
+          console.warn(
+            `Submitter ${submitterAddress} has an insufficient balance for the call and is likely to fail. Balance:`,
+            submitterBalance,
+            'Balance required:',
+            call.value,
+          );
+        }
+      }
+
       try {
+        if (
+          additionalTxSuccessCriteria &&
+          !additionalTxSuccessCriteria(chain, submitterAddress)
+        ) {
+          return false;
+        }
+        // Will throw if the transaction fails
         await multiProvider.estimateGas(chain, call, submitterAddress);
         return true;
       } catch (e) {} // eslint-disable-line no-empty
@@ -289,7 +351,11 @@ export abstract class HyperlaneAppGovernor<
     };
 
     if (await transactionSucceedsFromSender(chain, signerAddress)) {
-      return SubmissionType.SIGNER;
+      return {
+        type: SubmissionType.SIGNER,
+        chain,
+        call,
+      };
     }
 
     // 2. Check if the call will succeed via Gnosis Safe.
@@ -319,7 +385,11 @@ export abstract class HyperlaneAppGovernor<
               ))
           ) {
             console.warn(`${error.message}: Setting submission type to MANUAL`);
-            return SubmissionType.MANUAL;
+            return {
+              type: SubmissionType.MANUAL,
+              chain,
+              call,
+            };
           } else {
             console.error(
               `Failed to determine if signer can propose safe transactions: ${error}`,
@@ -333,11 +403,19 @@ export abstract class HyperlaneAppGovernor<
         this.canPropose[chain].get(safeAddress) &&
         (await transactionSucceedsFromSender(chain, safeAddress))
       ) {
-        return SubmissionType.SAFE;
+        return {
+          type: SubmissionType.SAFE,
+          chain,
+          call,
+        };
       }
     }
 
-    return SubmissionType.MANUAL;
+    return {
+      type: SubmissionType.MANUAL,
+      chain,
+      call,
+    };
   }
 
   handleOwnerViolation(violation: OwnerViolation) {