+++ typescript/sdk/src/core/HyperlaneCore.ts
@@ -60,6 +60,9 @@ export class HyperlaneCore extends HyperlaneApp<CoreFactories> {
       (chain, contracts): RouterConfig => ({
         mailbox: contracts.mailbox.address,
         owner: typeof owners === 'string' ? owners : owners[chain].owner,
+        ownerOverrides:
+          typeof owners === 'string' ? undefined : owners[chain].ownerOverrides,
+        proxyAdmin: contracts.proxyAdmin.address,
       }),
     );
     // filter for EVM chains