+++ typescript/sdk/src/ism/EvmIsmReader.ts
@@ -1,9 +1,11 @@
 import { ethers } from 'ethers';
 
 import {
+  ArbL2ToL1Ism__factory,
   DefaultFallbackRoutingIsm__factory,
   IInterchainSecurityModule__factory,
   IMultisigIsm__factory,
+  IOutbox__factory,
   OPStackIsm__factory,
   PausableIsm__factory,
   StaticAggregationIsm__factory,
@@ -25,6 +27,7 @@ import { HyperlaneReader } from '../utils/HyperlaneReader.js';
 
 import {
   AggregationIsmConfig,
+  ArbL2ToL1IsmConfig,
   IsmConfig,
   IsmType,
   ModuleType,
@@ -45,6 +48,9 @@ export interface IsmReader {
     address: Address,
   ): Promise<WithAddress<MultisigIsmConfig>>;
   deriveNullConfig(address: Address): Promise<WithAddress<NullIsmConfig>>;
+  deriveArbL2ToL1Config(
+    address: Address,
+  ): Promise<WithAddress<ArbL2ToL1IsmConfig>>;
   assertModuleType(
     moduleType: ModuleType,
     expectedModuleType: ModuleType,
@@ -100,6 +106,8 @@ export class EvmIsmReader extends HyperlaneReader implements IsmReader {
           break;
         case ModuleType.CCIP_READ:
           throw new Error('CCIP_READ does not have a corresponding IsmType');
+        case ModuleType.ARB_L2_TO_L1:
+          return this.deriveArbL2ToL1Config(address);
         default:
           throw new Error(`Unknown ISM ModuleType: ${moduleType}`);
       }
@@ -279,6 +287,21 @@ export class EvmIsmReader extends HyperlaneReader implements IsmReader {
     return {
       address,
       type: IsmType.TEST_ISM,
+    };
+  }
+
+  async deriveArbL2ToL1Config(
+    address: Address,
+  ): Promise<WithAddress<ArbL2ToL1IsmConfig>> {
+    const ism = ArbL2ToL1Ism__factory.connect(address, this.provider);
+
+    const outbox = await ism.arbOutbox();
+    const outboxContract = IOutbox__factory.connect(outbox, this.provider);
+    const bridge = await outboxContract.bridge();
+    return {
+      address,
+      type: IsmType.ARB_L2_TO_L1,
+      bridge,
     };
   }