+++ typescript/sdk/src/providers/SmartProvider/SmartProvider.ts
@@ -30,12 +30,23 @@ import {
 export const getSmartProviderErrorMessage = (errorMsg: string) =>
   `${errorMsg}: RPC request failed. Check RPC validity. To override RPC URLs, see: https://docs.hyperlane.xyz/docs/deploy-hyperlane-troubleshooting#override-rpc-urls`;
 
-// This is a partial list. If needed, check the full list for more: https://github.com/ethers-io/ethers.js/blob/fc66b8ad405df9e703d42a4b23bc452ec3be118f/src.ts/utils/errors.ts#L77-L85
+// This is a partial list. If needed, check the full list for more: https://docs.ethers.org/v5/api/utils/logger/#errors
 const RPC_SERVER_ERRORS = [
+  EthersError.NETWORK_ERROR,
   EthersError.NOT_IMPLEMENTED,
   EthersError.SERVER_ERROR,
+  EthersError.TIMEOUT,
   EthersError.UNKNOWN_ERROR,
   EthersError.UNSUPPORTED_OPERATION,
+];
+
+const RPC_BLOCKCHAIN_ERRORS = [
+  EthersError.CALL_EXCEPTION,
+  EthersError.INSUFFICIENT_FUNDS,
+  EthersError.NONCE_EXPIRED,
+  EthersError.REPLACEMENT_UNDERPRICED,
+  EthersError.TRANSACTION_REPLACED,
+  EthersError.UNPREDICTABLE_GAS_LIMIT,
 ];
 const DEFAULT_MAX_RETRIES = 1;
 const DEFAULT_BASE_RETRY_DELAY_MS = 250; // 0.25 seconds
@@ -424,15 +435,29 @@ export class HyperlaneSmartProvider
     const rpcServerError = errors.find((e) =>
       RPC_SERVER_ERRORS.includes(e.code),
     );
+
     const timedOutError = errors.find(
       (e) => e.status === ProviderStatus.Timeout,
     );
+
+    const rpcBlockchainError = errors.find((e) =>
+      RPC_BLOCKCHAIN_ERRORS.includes(e.code),
+    );
+
     if (rpcServerError) {
-      throw Error(getSmartProviderErrorMessage(rpcServerError.code));
+      throw Error(
+        rpcServerError.error?.message ?? // Server errors sometimes will not have an error.message
+          getSmartProviderErrorMessage(rpcServerError.code),
+      );
     } else if (timedOutError) {
       throw Error(getSmartProviderErrorMessage(ProviderStatus.Timeout));
+    } else if (rpcBlockchainError) {
+      throw Error(rpcBlockchainError.reason ?? rpcBlockchainError.code);
     } else {
-      throw Error(getSmartProviderErrorMessage(ProviderStatus.Error)); // Assumes that all errors are of ProviderStatus.Error
+      this.logger.error(
+        'Unhandled error case in combined provider error handler',
+      );
+      throw Error(fallbackMsg);
     }
   }
 }